requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
wechatpy>=1.8.0
httpx>=0.24.0
tenacity>=8.2.0
pybreaker>=1.0.1
slowapi>=0.1.9
structlog>=23.2.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

